requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-bootstrap-form==3.4
django-bootstrap4==3.0.1
django-ckeditor==6.2.0
django-filter==21.1
django-import-export==2.7.1
django-js-asset==1.2.2
djangorestframework==3.13.1
et-xmlfile==1.1.0
fontawesome-free==5.15.4
importlib-metadata==4.10.0
Markdown==3.3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.1.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
